fix(react): weird notification alignment
