+++ packages/panda-preset/src/recipes/slots/notification.ts
@@ -73,6 +73,7 @@ export const notification: Partial<SlotRecipeConfig> = defineSlotRecipe({
       opacity: '0',
       rounded: 'sm',
       shadow: 'md',
+      textAlign: 'left',
       _motionSafe: {
         animationName: 'fadeInDown',
         animationDuration: '250ms',
